# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NCS Warehouse Management API",
        "description": "API untuk manajemen gudang PT. NCS Bandung dengan algoritma Simulated Annealing",
        "contact": {
            "name": "NCS Development Team",
            "email": "dev@ncs.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
        "/api/area-gudang": {
            "get": {
                "tags": [
                    "Area Gudang"
                ],
                "summary": "Mendapatkan daftar area gudang",
                "description": "Mengambil daftar semua area gudang dengan pagination dan filter",
                "operationId": "getAreaGudangList",
                "parameters": [
                    {
                        "name": "gudang_id",
                        "in": "query",
                        "description": "Filter berdasarkan ID gudang",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "jenis_area",
                        "in": "query",
                        "description": "Filter berdasarkan jenis area (rak, lantai, khusus)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "rak",
                                "lantai",
                                "khusus"
                            ]
                        }
                    },
                    {
                        "name": "tersedia",
                        "in": "query",
                        "description": "Filter berdasarkan ketersediaan (1=tersedia, 0=tidak tersedia)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Pencarian berdasarkan nama area atau kode area",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman (default: 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar area gudang berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar area gudang berhasil diambil"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/AreaGudang"
                                                    }
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Area Gudang"
                ],
                "summary": "Membuat area gudang baru",
                "description": "Membuat data area gudang baru dengan validasi",
                "operationId": "createAreaGudang",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "gudang_id",
                                    "kode_area",
                                    "nama_area",
                                    "koordinat_x",
                                    "koordinat_y",
                                    "panjang",
                                    "lebar",
                                    "tinggi",
                                    "kapasitas",
                                    "jenis_area"
                                ],
                                "properties": {
                                    "gudang_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "kode_area": {
                                        "type": "string",
                                        "example": "A1-03"
                                    },
                                    "nama_area": {
                                        "type": "string",
                                        "example": "Area A1 - Rak Rendah"
                                    },
                                    "koordinat_x": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25
                                    },
                                    "koordinat_y": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5
                                    },
                                    "panjang": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "lebar": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 8
                                    },
                                    "tinggi": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3
                                    },
                                    "kapasitas": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 240
                                    },
                                    "jenis_area": {
                                        "type": "string",
                                        "enum": [
                                            "rak",
                                            "lantai",
                                            "khusus"
                                        ],
                                        "example": "rak"
                                    },
                                    "tersedia": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Area gudang berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Area gudang berhasil dibuat"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AreaGudang"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/area-gudang/{id}/toggle-status": {
            "patch": {
                "tags": [
                    "Area Gudang"
                ],
                "summary": "Toggle status tersedia area gudang",
                "description": "Mengubah status tersedia/tidak tersedia area gudang",
                "operationId": "toggleAreaGudangStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Area Gudang",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status area gudang berhasil diubah",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Status area gudang berhasil diubah"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AreaGudang"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Area gudang tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}